+++ vigilant_crypto_snatch/commands/watch.py
@@ -1,5 +1,6 @@
 import datetime
 
+from .. import logger
 from ..configuration import get_used_currencies
 from ..configuration import run_migrations
 from ..configuration import YamlConfiguration
@@ -17,20 +18,18 @@ from ..triggers import make_triggers
 from ..watchloop import TriggerLoop
 
 
-def main(marketplace_name, keepalive, one_shot, dry_run):
+def main(marketplace_name):
     run_migrations()
     config = YamlConfiguration()
 
     add_telegram_logger(config.get_telegram_config())
-    if not one_shot:
-        logger.info("Starting up …")
+    logger.info("Starting up …")
 
     datastore = make_datastore(user_db_path)
-    market = make_marketplace(marketplace_name, config, dry_run)
+    market = make_marketplace(marketplace_name, config)
     check_and_perform_widthdrawal(market)
 
-    if not one_shot:
-        report_balances(market, get_used_currencies(config.get_trigger_config()))
+    report_balances(market, get_used_currencies(config.get_trigger_config()))
 
     database_source = DatabaseHistoricalSource(datastore, datetime.timedelta(minutes=5))
     crypto_compare_source = CryptoCompareHistoricalSource(
@@ -41,10 +40,8 @@ def main(marketplace_name, keepalive, one_shot, dry_run):
         database_source, [market_source, crypto_compare_source], datastore
     )
     active_triggers = make_triggers(
-        config.get_trigger_config(), datastore, caching_source, market, dry_run
+        config.get_trigger_config(), datastore, caching_source, market
     )
 
-    trigger_loop = TriggerLoop(
-        active_triggers, config.get_polling_interval(), keepalive, one_shot
-    )
+    trigger_loop = TriggerLoop(active_triggers, config.get_polling_interval())
     trigger_loop.loop()